otel/requirements.txt
azure-core==1.34.0
azure-identity==1.23.0
azure-monitor-opentelemetry-exporter==1.0.0b37
blacksheep==2.3.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
essentials==1.1.6
essentials-openapi==1.2.0
fixedint==0.1.6
googleapis-common-protos==1.70.0
grpcio==1.72.1
guardpost==1.0.2
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.34.0
opentelemetry-distro==0.55b0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asyncio==0.55b0
opentelemetry-instrumentation-dbapi==0.55b0
opentelemetry-instrumentation-grpc==0.55b0
opentelemetry-instrumentation-logging==0.55b0
opentelemetry-instrumentation-requests==0.55b0
opentelemetry-instrumentation-sqlite3==0.55b0
opentelemetry-instrumentation-threading==0.55b0
opentelemetry-instrumentation-urllib==0.55b0
opentelemetry-instrumentation-urllib3==0.55b0
opentelemetry-instrumentation-wsgi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
packaging==25.0
protobuf==5.29.5
psutil==7.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rodi==2.0.8
six==1.17.0
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
zipp==3.23.0
